refactor(frontend): collapse F2P, items and leaderboard duplication
Steps 2, 5 and 9 of docs/frontend-unification-design.md. Net -4,121 lines.

Step 2 - one F2P module (-1,002)

lib/src/f2p/* and lib/src/f2p-evm/* become lib/src/fights/f2p.js,
parameterised by {chainid, credential} rather than forked per chain. The
HTML ids were normalised (-f2p-evm to -f2p, -selecter to -selected).

The doc understated this one: those ids are also referenced by 8 CSS
files, lib/custom.js and lib/src/openFights.js, and -selecter is used by
four NON-F2P lines in lib/index.js driving P2P game creation. Sed-ing only
the HTML, as the doc suggested, would have broken EVM P2P create-game.

Both regressions the EVM copy had dropped are restored: the error-modal
branches (EVM previously swallowed a failed withdraw/join silently) and
the "Invite a friend" link.

Bugs fixed in passing: fights.forEach(async ...) ran the emptiness check
before rows were appended, which is why the list was rendered twice; a
re-render after withdraw sent sign_evm undefined and threw; a live
NodeList was mutated during iteration; desktop tab listeners were re-bound
on every 7.5s poll.

Step 5 - one item pipeline (-2,849, beating the doc's -2,375)

Six files (5,066 lines) become five shared modules under lib/src/items/
(2,175) plus thin facades where lib/index*.js still imports the old paths.
The three genuine EVM/TVM differences are handled as the doc specifies:
NFT enumeration behind one adapter method, per-item nftAddress kept as one
optional field gated on a capability so EVM request bodies and DOM stay
byte-identical, and the endpoint prefix as chain.route().

The ton/lootbox -> EVM inventory cross-import is gone by construction:
main_ton.js drops 1.24 MiB to 954 KiB.

Step 9 - one leaderboard (-223)

The doc said only the TON leaderboard was mock data. Both were: the EVM
module also had its fetch commented out and was rendering hardcoded
chainid-31337 Hardhat rows to real users. Deleting only the TON mock would
have shown EVM mock data on the TON page, so the fetch is live on both.
/leaderboard/ton already existed and returns real data; only the client
call was missing.

Verified: clean webpack build, 0 errors. Both lobbies render in a real
browser - EVM on desktop, the TON Mini App on mobile (its "platform not
supported" gate is by design, not a regression). Server suites green:
validation, sign-in nonce, socket auth 5/5.

Not verified: no wallet, so buying an NFT, opening a lootbox and
drag-equipping were traced through the merged code rather than executed.

+++ server/leaderboard/service.js
@@ -167,9 +167,9 @@ export async function createLeaderboardTON(req, res) {
                 player,
                 SUM(kills) AS total_kills,
                 SUM(deaths) AS total_deaths,
-                COUNT(*) FILTER (WHERE baseAmount < amount) AS wins,
+                COUNT(*) FILTER (WHERE baseAmount::DECIMAL < amount::DECIMAL) AS wins,
                 COUNT(*) AS total_games,
-                SUM(CASE WHEN baseAmount < amount THEN (amount::DECIMAL - baseAmount::DECIMAL) ELSE 0 END) / 10^9 * $2 AS amount_won
+                SUM(CASE WHEN baseAmount::DECIMAL < amount::DECIMAL THEN (amount::DECIMAL - baseAmount::DECIMAL) ELSE 0 END) / 10^9 * $2 AS amount_won
             FROM
                 statistics
             WHERE
@@ -185,9 +185,9 @@ export async function createLeaderboardTON(req, res) {
                 player,
                 SUM(kills) AS total_kills,
                 SUM(deaths) AS total_deaths,
-                COUNT(*) FILTER (WHERE baseAmount < amount) AS wins,
+                COUNT(*) FILTER (WHERE baseAmount::DECIMAL < amount::DECIMAL) AS wins,
                 COUNT(*) AS total_games,
-                SUM(CASE WHEN baseAmount < amount THEN (amount::DECIMAL - baseAmount::DECIMAL) ELSE 0 END) / 10^9 * $2 AS amount_won
+                SUM(CASE WHEN baseAmount::DECIMAL < amount::DECIMAL THEN (amount::DECIMAL - baseAmount::DECIMAL) ELSE 0 END) / 10^9 * $2 AS amount_won
             FROM
                 statistics
             WHERE
